+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+- audio icon on sidebar tab in system theme
+
 ## [0.0.11] - 2019-07-07
 
 ### Changed